NAS-130211: Rename network widget to interface widget
