+++ src/app/pages/dashboard/widgets/network/widget-interface/widget-interface.component.html
@@ -2,7 +2,7 @@
   <mat-card-content class="card-content">
     @if (isFullSize()) {
       <div class="header">
-        <h3>{{ 'Network' | translate }}</h3>
+        <h3>{{ 'Interface' | translate }}</h3>
         <button
           mat-icon-button
           matTooltipPosition="above"
