+++ src/app/pages/dashboard/widgets/network/widget-interface/widget-interface.const.ts
