+++ src/app/pages/dashboard/widgets/network/widget-interface/widget-interface.utils.ts
